second commit

+++ src/main/java/com/example/Gestiondesnotes/services/ProfDAO.java
@@ -2,6 +2,8 @@ package com.example.Gestiondesnotes.services;
 
 
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,22 +18,28 @@ public class ProfDAO {
 	public void save(Prof prof) {
 		profrepository.save(prof);
 	}
-	public void delete(int cinprof) {
-		profrepository.deleteById(cinprof);
-	}
+	 public void delete(int cinprof) {
+		 if (!profrepository.existsById(cinprof)) {
+			 throw new RuntimeException("Professor not found");
+			 }
+		 profrepository.deleteById(cinprof);
+	 }
 	public Prof getProfByEmail(String email) {
         return profrepository.findByEmail(email);
     }
+	public Prof getProfById(int cin) {
+		return profrepository.findById(cin).orElse(null);
+	}
 	 public boolean authenticateProf(String email, String password) {
 		 Prof prof = getProfByEmail(email);
 	     if (prof == null) {
 	    	 return false; // User not found
-	     }
+	 }
 	     return prof.getPassword().equals(password); // Simple password check
 	 }
 	 public Prof addProf(int cin,String nom,String prenom,String gener, String email,String password) {
 	        // Check if email already exists
-	        if (getProfByEmail(email) != null) {
+	        if (getProfByEmail(email) != null || getProfById(cin) != null) {
 	            throw new IllegalArgumentException("Email is already registered");
 	        }
 
@@ -45,6 +53,9 @@ public class ProfDAO {
 	        newProf.setPassword(password);
 	        
 	        return profrepository.save(newProf);
+	    }
+	 public List<Prof> getAllProfessors() {
+	        return profrepository.findAll();
 	    }
 
 }
